Move classlist_exists funct to class_registry.py, add import.
